REFACTOR: generalizing patterns
