+++ src/patterns/stripe.c
@@ -3,9 +3,12 @@
 t_pattern stripe_pattern(t_color color_a, t_color color_b)
 {
 	return ((t_pattern){
+		.pattern_type = STRIPE,
 		.has_pattern = true,
 		.color_a = color_a,
 		.color_b = color_b,
+		.transform = identity(),
+		.inverse = identity()
 	});
 }
 
@@ -36,23 +39,4 @@ t_color	stripe_at(t_pattern pattern, t_point pattern_point)
 	else
 		new_color = pattern.color_b;
 	return (new_color);
-}
-
-t_color	stripe_at_object(t_pattern pattern, t_shape object, t_point world_point)
-{
-	t_point	object_point;
-	t_point	pattern_point;
-	t_color	color_stripe;
-
-	object_point = mx_multiply_tuple(object.inverse, world_point);
-	pattern_point = mx_multiply_tuple(pattern.inverse, object_point);
-	color_stripe = stripe_at(pattern, pattern_point);
-
-	return (color_stripe);
-}
-
-void	set_pattern_transformation(t_pattern *pattern, t_matrix transformation)
-{
-	pattern->transform = transformation;
-	pattern->inverse = inverse(transformation);
 }
